+++ packages/article/__tests__/related-articles-shared.js
@@ -1,8 +1,7 @@
-/* global context */
-
 import "react-native";
 import React from "react";
 import renderer from "react-test-renderer";
+import mockDate from "mockdate";
 
 import RelatedArticles from "../related-articles/related-articles";
 
@@ -13,16 +12,18 @@ import singleRelatedArticleNoBylineFixture from "../related-articles/fixtures/si
 import twoRelatedArticlesFixture from "../related-articles/fixtures/two-related-articles.json";
 import threeRelatedArticlesFixture from "../related-articles/fixtures/three-related-articles.json";
 
-const createRelatedArticlesProps = fixtureData => ({
+const createRelatedArticlesProps = (fixtureData, action = () => {}) => ({
+  analyticsStream: action,
   articles: fixtureData.relatedArticles,
   template: fixtureData.relatedArticlesLayout.template,
   onPress: () => {}
 });
 
-module.exports = () => {
+export default () => {
   const realIntl = Intl;
 
   beforeEach(() => {
+    mockDate.set("1/1/2018");
     global.Intl = {
       DateTimeFormat: () => ({
         resolvedOptions: () => ({ timeZone: "Europe/London" })
@@ -31,10 +32,11 @@ module.exports = () => {
   });
 
   afterEach(() => {
+    mockDate.reset();
     global.Intl = realIntl;
   });
 
-  context("Related articles", () => {
+  describe("Related articles", () => {
     it("handles no related articles", () => {
       const data = {
         relatedArticles: [],
@@ -118,6 +120,74 @@ module.exports = () => {
         )
         .toJSON();
       expect(tree).toMatchSnapshot();
+    });
+  });
+
+  describe("tracker", () => {
+    it("sends analytics even if no related articles are rendered", () => {
+      const events = jest.fn();
+      const data = {
+        relatedArticles: [],
+        relatedArticlesLayout: {
+          template: "DEFAULT"
+        }
+      };
+      renderer.create(
+        <RelatedArticles {...createRelatedArticlesProps(data, events)} />
+      );
+      expect(events.mock.calls).toMatchSnapshot();
+    });
+
+    it("sends analytics if a single related article is rendered", () => {
+      const events = jest.fn();
+      renderer.create(
+        <RelatedArticles
+          {...createRelatedArticlesProps(
+            singleRelatedArticleFixture.data,
+            events
+          )}
+        />
+      );
+      expect(events.mock.calls).toMatchSnapshot();
+    });
+
+    it("sends analytics if a single related article with no byline is rendered", () => {
+      const events = jest.fn();
+      renderer.create(
+        <RelatedArticles
+          {...createRelatedArticlesProps(
+            singleRelatedArticleNoBylineFixture.data,
+            events
+          )}
+        />
+      );
+      expect(events.mock.calls).toMatchSnapshot();
+    });
+
+    it("sends analytics if two related articles are rendered", () => {
+      const events = jest.fn();
+      renderer.create(
+        <RelatedArticles
+          {...createRelatedArticlesProps(
+            twoRelatedArticlesFixture.data,
+            events
+          )}
+        />
+      );
+      expect(events.mock.calls).toMatchSnapshot();
+    });
+
+    it("sends analytics if three related articles are rendered", () => {
+      const events = jest.fn();
+      renderer.create(
+        <RelatedArticles
+          {...createRelatedArticlesProps(
+            threeRelatedArticlesFixture.data,
+            events
+          )}
+        />
+      );
+      expect(events.mock.calls).toMatchSnapshot();
     });
   });
 };