+++ packages/article/styles/article-header/responsive.web.js
@@ -1,10 +1,11 @@
 import { View, Text } from "react-native";
 import withResponsiveStyles from "@times-components/responsive-styles";
+import { colours } from "@times-components/styleguide";
 
 export const HeadlineContainer = withResponsiveStyles(Text, {
   base: () => `
     font-size: 30px;
-    color: #1d1d1b;
+    color: ${colours.functional.timesBlack};
     margin-bottom: 8px;
     font-family: "TimesModern-Bold";
     font-weight: 400;