+++ packages/card/__tests__/android/card.android.test.js
@@ -1,5 +1,5 @@
 import cardTests from "../card.native.test.js";
 
-describe("Card test on android", () => {
+describe("Card tests on android", () => {
   cardTests();
 });