+++ packages/icons/icons.js
@@ -1,5 +1,7 @@
 import diamond from "./icons/diamond";
 import twitter from "./icons/twitter";
+import video from "./icons/video";
 
 export const IconDiamond = diamond;
 export const IconTwitter = twitter;
+export const IconVideo = video;