+++ packages/icons/icons/diamond.js
@@ -1,7 +1,7 @@
 import React from "react";
-import PropTypes from "prop-types";
-
+import { colours } from "@times-components/styleguide";
 import Svg, { G, Path } from "svgs";
+import iconPropTypes from "./proptypes";
 
 const IconDiamond = ({ width, height, fillColour }) => (
   <Svg width={width} height={height} viewBox="0 0 20 20">
@@ -11,10 +11,10 @@ const IconDiamond = ({ width, height, fillColour }) => (
   </Svg>
 );
 
-IconDiamond.propTypes = {
-  width: PropTypes.number.isRequired,
-  height: PropTypes.number.isRequired,
-  fillColour: PropTypes.string.isRequired
+IconDiamond.propTypes = iconPropTypes;
+
+IconDiamond.defaultProps = {
+  fillColour: colours.functional.timesBlack
 };
 
 export default IconDiamond;