+++ packages/slice/index.js
@@ -1,6 +1,5 @@
-// @TODO: delete this file and move this code to slice.js when related articles wired up properly
 import Slice from "./slice";
-import DefaultSlice from "./templates/default/default";
-import LeadSlice from "./templates/lead/lead";
+import DefaultSlice from "./templates/default";
+import LeadSlice from "./templates/lead";
 
 export { Slice, DefaultSlice, LeadSlice };