+++ packages/article-byline/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.12.15"></a>
+## [0.12.15](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.12.14...@times-components/article-byline@0.12.15) (2018-03-06)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-byline
+
+<a name="0.12.14"></a>
+## [0.12.14](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.12.13...@times-components/article-byline@0.12.14) (2018-03-06)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-byline
+
+<a name="0.12.13"></a>
+## [0.12.13](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.12.12...@times-components/article-byline@0.12.13) (2018-03-02)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-byline
+
+<a name="0.12.12"></a>
+## [0.12.12](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.12.11...@times-components/article-byline@0.12.12) (2018-03-01)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-byline
+
 <a name="0.12.11"></a>
 ## [0.12.11](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.12.10...@times-components/article-byline@0.12.11) (2018-02-28)