+++ packages/author-head/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.19.11"></a>
+## [0.19.11](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.19.10...@times-components/author-head@0.19.11) (2018-03-06)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-head
+
+<a name="0.19.10"></a>
+## [0.19.10](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.19.9...@times-components/author-head@0.19.10) (2018-03-06)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-head
+
+<a name="0.19.9"></a>
+## [0.19.9](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.19.8...@times-components/author-head@0.19.9) (2018-03-02)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-head
+
+<a name="0.19.8"></a>
+## [0.19.8](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.19.6...@times-components/author-head@0.19.8) (2018-03-01)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-head
+
+<a name="0.19.7"></a>
+## [0.19.7](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.19.6...@times-components/author-head@0.19.7) (2018-03-01)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-head
+
+<a name="0.19.6"></a>
+## [0.19.6](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.19.5...@times-components/author-head@0.19.6) (2018-03-01)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-head
+
+<a name="0.19.5"></a>
+## [0.19.5](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.19.4...@times-components/author-head@0.19.5) (2018-02-28)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-head
+
 <a name="0.19.4"></a>
 ## [0.19.4](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.19.3...@times-components/author-head@0.19.4) (2018-02-28)