+++ packages/markup/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.21.11"></a>
+## [0.21.11](https://github.com/newsuk/times-components/compare/@times-components/markup@0.21.10...@times-components/markup@0.21.11) (2018-03-06)
+
+
+
+
+**Note:** Version bump only for package @times-components/markup
+
+<a name="0.21.10"></a>
+## [0.21.10](https://github.com/newsuk/times-components/compare/@times-components/markup@0.21.9...@times-components/markup@0.21.10) (2018-03-06)
+
+
+
+
+**Note:** Version bump only for package @times-components/markup
+
+<a name="0.21.9"></a>
+## [0.21.9](https://github.com/newsuk/times-components/compare/@times-components/markup@0.21.8...@times-components/markup@0.21.9) (2018-03-02)
+
+
+
+
+**Note:** Version bump only for package @times-components/markup
+
+<a name="0.21.8"></a>
+## [0.21.8](https://github.com/newsuk/times-components/compare/@times-components/markup@0.21.7...@times-components/markup@0.21.8) (2018-03-01)
+
+
+
+
+**Note:** Version bump only for package @times-components/markup
+
 <a name="0.21.7"></a>
 ## [0.21.7](https://github.com/newsuk/times-components/compare/@times-components/markup@0.21.6...@times-components/markup@0.21.7) (2018-02-28)